backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
validators==0.35.0

# Scheduler for reminders
apscheduler==3.10.4

# Push notifications (Firebase)
firebase-admin==6.3.0

# HTTP requests (for notifications)
httpx==0.25.2

# Development/Testing
pytest==7.4.3
pytest-asyncio==0.21.1